+++ packages/expo-dev-launcher/ios/ReactDelegateHandler/ExpoDevLauncherAppDelegateSubscriber.swift
@@ -4,10 +4,6 @@ import ExpoModulesCore
 
 public class ExpoDevLauncherAppDelegateSubscriber: ExpoAppDelegateSubscriber {
   public func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
-    if !ExpoDevLauncherReactDelegateHandler.shouldEnableAutoSetup {
-      return false
-    }
-
     guard let window = UIApplication.shared.delegate?.window ?? UIApplication.shared.windows.filter { $0.isKeyWindow }.first else {
       fatalError("Cannot find the keyWindow. Make sure to call `window.makeKeyAndVisible()`.")
     }
@@ -16,10 +12,6 @@ public class ExpoDevLauncherAppDelegateSubscriber: ExpoAppDelegateSubscriber {
   }
 
   public func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
-    if !ExpoDevLauncherReactDelegateHandler.shouldEnableAutoSetup {
-      return false
-    }
-
     return EXDevLauncherController.sharedInstance().onDeepLink(url, options: options)
   }
 }